Modifications to Parameter Input
Modifications to denssopts.py and saxstats that allow for individual parameters to be entered for each scattering profile. Other edits include the ability to average the profiles together after 'avg_steps' number of fitting steps, which provides the opportunity for the code to change how it fits data. The support was also changed into an array, so that each reconstruction has its own support. The support of the individual reconstruction is never influenced by the others, even when the density is averaged together.