Scaling with actual neutron scattering length densities
I changed 'sld', which used to represent the buffer scattering length density, to 'blsd', and added 'tsld', which is the SLD of the target you are observing in solution (such as a protein) Realistically, any observed SLDs will be in the range between the buffer and the target. You can enter in the 'tsld' values on the command line or using the -paramf (parameter file). Testing of the 'tsld' linear scaling is still in progress, so expect further commits revising any bugs.